requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart